CR updates
