+++ components/rewards/hooks/useRewards.tsx
@@ -8,7 +8,8 @@ import type { RewardStatusFilter } from 'components/rewards/components/RewardVie
 import { useCurrentSpace } from 'hooks/useCurrentSpace';
 import { usePages } from 'hooks/usePages';
 import { useUser } from 'hooks/useUser';
-import type { RewardCreationData, Reward, RewardWithUsers } from 'lib/rewards/interfaces';
+import type { RewardCreationData } from 'lib/rewards/createReward';
+import type { RewardWithUsers } from 'lib/rewards/interfaces';
 import type { RewardUpdate } from 'lib/rewards/updateRewardSettings';
 
 type RewardsContextType = {
