+++ lib/rewards/interfaces.ts
@@ -2,6 +2,8 @@ import type { TargetPermissionGroup } from '@charmverse/core/permissions';
 import type { Application, Bounty, BountyStatus, Page, Transaction } from '@charmverse/core/prisma-client';
 
 import type { PageContent } from 'lib/prosemirror/interfaces';
+import type { RewardPageProps } from 'lib/rewards/createReward';
+import type { UpdateableRewardFields } from 'lib/rewards/updateRewardSettings';
 
 export type RewardReviewer = TargetPermissionGroup<'role' | 'user'>;
 
@@ -29,32 +31,6 @@ export type RewardWithUsersAndPageMeta = Bounty & {
 
 export type ApplicationWithTransactions = Application & {
   transactions: Transaction[];
-};
-
-export type UpdateableRewardFields = Partial<
-  Pick<
-    Reward,
-    | 'chainId'
-    | 'rewardAmount'
-    | 'rewardToken'
-    | 'approveSubmitters'
-    | 'allowMultipleApplications'
-    | 'maxSubmissions'
-    | 'dueDate'
-    | 'customReward'
-    | 'fields'
-  >
-> & { reviewers?: RewardReviewer[]; allowedSubmitterRoles?: string[] | null };
-
-export type RewardPageProps = Partial<
-  Pick<Page, 'title' | 'content' | 'contentText' | 'sourceTemplateId' | 'headerImage' | 'icon'>
->;
-
-export type RewardCreationData = UpdateableRewardFields & {
-  pageProps?: RewardPageProps;
-  linkedPageId?: string;
-  spaceId: string;
-  userId: string;
 };
 
 export type RewardPageAndPropertiesInput = RewardPageProps & UpdateableRewardFields & { content?: PageContent | null };
