+++ pages/api/rewards/[id]/index.ts
@@ -5,8 +5,9 @@ import nc from 'next-connect';
 import { onError, onNoMatch, requireUser } from 'lib/middleware';
 import { providePermissionClients } from 'lib/permissions/api/permissionsClientMiddleware';
 import { getRewardOrThrow } from 'lib/rewards/getReward';
-import type { RewardWithUsers, RewardWithUsersAndPageMeta, UpdateableRewardFields } from 'lib/rewards/interfaces';
+import type { RewardWithUsers, RewardWithUsersAndPageMeta } from 'lib/rewards/interfaces';
 import { rollupRewardStatus } from 'lib/rewards/rollupRewardStatus';
+import type { UpdateableRewardFields } from 'lib/rewards/updateRewardSettings';
 import { updateRewardSettings } from 'lib/rewards/updateRewardSettings';
 import { withSessionRoute } from 'lib/session/withSession';
 import { UnauthorisedActionError } from 'lib/utilities/errors';
